build(main): Move traefik to proper profile
Switch from local override strategy to custom cloud override that would allow us to work locally by default, and use traefik and other production configs only when needed.
Since this is a course project, we'll most use it locally. Preserving "cloud" config is only for reference, in case we end up reaching some staging/production state or for future references on folks looking on how to do it.